requirements.txt
torch==1.8.1
numpy==1.18.1
transformers==4.5.0
tqdm==4.42.1
seqeval==0.0.3
pytorch_crf==0.7.2
